+++ compiled/src/mixin-utilities.d.ts
@@ -1,2 +1,0 @@
-export declare function applyMixins(derivedCtor: any, baseCtors: any[]): void;
-export declare function applyMixinsExcludeConstructor(derivedCtor: any, baseCtors: any[]): void;