+++ compiled/src/mixin.decorator.d.ts
@@ -1,1 +1,0 @@
-export declare function Mixin(classes: any[], ignoreConstructor?: boolean): (target: Function) => void;