+++ src/index.ts
@@ -1,3 +1,2 @@
 export * from './event-manager';
 export * from './event-listener.decorator';
-export * from './mixin.decorator';